+++ mjx/mujoco/mjx/warp/collision_driver.py
@@ -42,7 +42,6 @@ _e = mjwarp.Constraint(
     **{f.name: None for f in dataclasses.fields(mjwarp.Constraint) if f.init}
 )
 
-
 @ffi.format_args_for_warp
 def _collision_shim(
     # Model
@@ -279,6 +278,7 @@ def _collision_jax_impl(m: types.Model, d: types.Data):
           'contact__type',
           'contact__worldid',
       },
+      graph_mode=m.opt._impl.graph_mode,
   )
   out = jf(
       d.qpos.shape[0],